requirements.txt
Flask
Flask-SQLAlchemy
Flask-PyMongo
pymysql
requests
Flask-RESTful
PyJWT
jwt